Add Base Data Setup up process frmo CSV file

+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/ComponentAlternativeDAO.java
@@ -132,11 +132,11 @@ public interface ComponentAlternativeDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
 
     /*
      * Returns total amount of rows in table.
      */
-    public int countAll() throws Exception;
+    public long countAll() throws Exception;
 
 }
