+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/ComponentCommentDAO.java
@@ -113,11 +113,11 @@ public interface ComponentCommentDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
 
     /*
      * Returns total amount of rows in table.
      */
-    public int countAll() throws Exception;
+    public long countAll() throws Exception;
  
 }
