+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/ComponentRelationshipDAO.java
@@ -125,11 +125,11 @@ public interface ComponentRelationshipDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
 
     /*
      * Returns total amount of rows in table.
      */
-    public int countAll() throws Exception;
+    public long countAll() throws Exception;
 
 }
