+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/ComponentSynonymDAO.java
@@ -116,11 +116,11 @@ public interface ComponentSynonymDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
 
     /*
      * Returns total amount of rows in table.
      */
-    public int countAll() throws Exception;
+    public long countAll() throws Exception;
 
 }
