+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/DerivedPartOfDAO.java
@@ -108,6 +108,6 @@ public interface DerivedPartOfDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
 
 }
