+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/DerivedPartOfFKDAO.java
@@ -70,6 +70,6 @@ public interface DerivedPartOfFKDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchTerm) throws Exception;
+    public long count(String searchTerm) throws Exception;
 
 }
