+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/DerivedPartOfPerspectivesFKDAO.java
@@ -54,6 +54,6 @@ public interface DerivedPartOfPerspectivesFKDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchTerm, String searchId, String searchDirection, String searchStartStage, String searchEndStage, String searchPerspective) throws Exception;
+    public long count(String searchTerm, String searchId, String searchDirection, String searchStartStage, String searchEndStage, String searchPerspective) throws Exception;
 
 }
