+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/EditorDAO.java
@@ -90,6 +90,6 @@ public interface EditorDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchTerm) throws Exception;
+    public long count(String searchTerm) throws Exception;
 
 }
