+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/ExtraTimedNodeDAO.java
@@ -86,6 +86,6 @@ public interface ExtraTimedNodeDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
 
 }
