+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/JOINTimedNodeStageDAO.java
@@ -76,26 +76,26 @@ public interface JOINTimedNodeStageDAO extends BaseDAO {
     /*
      * Returns a count of ALL jointimednodestages by node FK, otherwise null.
      */
-    public int countAllByNodeFk(long nodeFk) throws Exception;
+    public long countAllByNodeFk(long nodeFk) throws Exception;
     
     /*
      * Returns a count of ALL jointimednodestages by node FK, otherwise null.
      */
-    public int maxSequenceByNodeFk(long nodeFk) throws Exception;
+    public long maxSequenceByNodeFk(long nodeFk) throws Exception;
     
     /*
      * Returns a count of ALL jointimednodestages by node FK, otherwise null.
      */
-    public int minSequenceByNodeFk(long nodeFk) throws Exception;
+    public long minSequenceByNodeFk(long nodeFk) throws Exception;
     
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String sql, long key) throws Exception;
+    public long count(String sql, long key) throws Exception;
     
     /*
      * Returns total amount of rows in table.
      */
-    public int maxMin(String sql, long key) throws Exception;
+    public long maxMin(String sql, long key) throws Exception;
     
 }
