+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/LogDAO.java
@@ -47,12 +47,12 @@ public interface LogDAO extends BaseDAO {
     /*
      * Returns the maximum Oid.
      */
-    public int maximumOid() throws Exception;
+    public long maximumOid() throws Exception;
     
     /*
      * Returns the maximum Logged Oid.
      */
-    public int maximumLoggedOid() throws Exception;
+    public long maximumLoggedOid() throws Exception;
     
     /*
      * Returns the log from the database matching the given OID, otherwise null.
@@ -118,11 +118,11 @@ public interface LogDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
     
     /*
      * Returns total amount of rows in table.
      */
-    public int maximum(String sql) throws Exception;
+    public long maximum(String sql) throws Exception;
     
 }
