+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/PerspectiveAmbitDAO.java
@@ -91,6 +91,6 @@ public interface PerspectiveAmbitDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchTerm, String searchExtra) throws Exception;
+    public long count(String searchTerm, String searchExtra) throws Exception;
 
 }
