+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/SourceFormatDAO.java
@@ -90,6 +90,6 @@ public interface SourceFormatDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchTerm, String searchExtra) throws Exception;
+    public long count(String searchTerm, String searchExtra) throws Exception;
 
 }
