+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/SpeciesDAO.java
@@ -102,6 +102,6 @@ public interface SpeciesDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
 
 }
