+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/SynonymDAO.java
@@ -112,6 +112,6 @@ public interface SynonymDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
 
 }
