+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daointerface/ThingDAO.java
@@ -51,7 +51,7 @@ public interface ThingDAO extends BaseDAO {
     /*
      * Returns the maximum Oid.
      */
-    public int maximumOid() throws Exception;
+    public long maximumOid() throws Exception;
     
     /*
      * Returns the Thing from the database matching the given OID, otherwise null.
@@ -112,11 +112,11 @@ public interface ThingDAO extends BaseDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception;
+    public long count(String searchFirst, String searchSecond) throws Exception;
 
     /*
      * Returns total amount of rows in table.
      */
-    public int maximum(String sql) throws Exception;
+    public long maximum(String sql) throws Exception;
     
 }
