+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daojdbc/ComponentAlternativeDAOJDBC.java
@@ -607,7 +607,7 @@ public final class ComponentAlternativeDAOJDBC implements ComponentAlternativeDA
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception {
+    public long count(String searchFirst, String searchSecond) throws Exception {
 
         String searchFirstWithWildCards = "";
         String searchSecondWithWildCards = "";
@@ -634,7 +634,7 @@ public final class ComponentAlternativeDAOJDBC implements ComponentAlternativeDA
         Connection connection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        int count = 0;
+        long count = 0;
 
         try {
         	
@@ -645,7 +645,7 @@ public final class ComponentAlternativeDAOJDBC implements ComponentAlternativeDA
 
             if (resultSet.next()) {
             	
-                count = resultSet.getInt("VALUE");
+                count = resultSet.getLong("VALUE");
             }
         } 
         catch (SQLException e) {
@@ -663,12 +663,12 @@ public final class ComponentAlternativeDAOJDBC implements ComponentAlternativeDA
     /*
      * Returns total amount of rows in table.
      */
-    public int countAll() throws Exception {
+    public long countAll() throws Exception {
 
         Connection connection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        int count = 0;
+        long count = 0;
 
         try {
         	
@@ -679,7 +679,7 @@ public final class ComponentAlternativeDAOJDBC implements ComponentAlternativeDA
 
             if (resultSet.next()) {
             	
-                count = resultSet.getInt("VALUE");
+                count = resultSet.getLong("VALUE");
             }
         } 
         catch (SQLException e) {
