+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daojdbc/ComponentCommentDAOJDBC.java
@@ -566,7 +566,7 @@ public final class ComponentCommentDAOJDBC implements ComponentCommentDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception {
+    public long count(String searchFirst, String searchSecond) throws Exception {
 
         String searchFirstWithWildCards = "";
         String searchSecondWithWildCards = "";
@@ -593,7 +593,7 @@ public final class ComponentCommentDAOJDBC implements ComponentCommentDAO {
         Connection connection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        int count = 0;
+        long count = 0;
 
         try {
         	
@@ -604,7 +604,7 @@ public final class ComponentCommentDAOJDBC implements ComponentCommentDAO {
 
             if (resultSet.next()) {
             	
-                count = resultSet.getInt("VALUE");
+                count = resultSet.getLong("VALUE");
             }
             
         } 
@@ -623,12 +623,12 @@ public final class ComponentCommentDAOJDBC implements ComponentCommentDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int countAll() throws Exception {
+    public long countAll() throws Exception {
 
         Connection connection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        int count = 0;
+        long count = 0;
 
         try {
         	
@@ -639,7 +639,7 @@ public final class ComponentCommentDAOJDBC implements ComponentCommentDAO {
 
             if (resultSet.next()) {
             	
-                count = resultSet.getInt("VALUE");
+                count = resultSet.getLong("VALUE");
             }
         } 
         catch (SQLException e) {
