+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daojdbc/ComponentOrderDAOJDBC.java
@@ -683,7 +683,7 @@ public final class ComponentOrderDAOJDBC implements ComponentOrderDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchFirst, String searchSecond) throws Exception {
+    public long count(String searchFirst, String searchSecond) throws Exception {
 
         String searchFirstWithWildCards = "";
         String searchSecondWithWildCards = "";
@@ -710,7 +710,7 @@ public final class ComponentOrderDAOJDBC implements ComponentOrderDAO {
         Connection connection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        int count = 0;
+        long count = 0;
 
         try {
         	
@@ -721,7 +721,7 @@ public final class ComponentOrderDAOJDBC implements ComponentOrderDAO {
 
             if (resultSet.next()) {
             	
-                count = resultSet.getInt("VALUE");
+                count = resultSet.getLong("VALUE");
             }
         } 
         catch (SQLException e) {
@@ -739,12 +739,12 @@ public final class ComponentOrderDAOJDBC implements ComponentOrderDAO {
     /*
      * Returns total amount of rows in table.
      */
-    public int countAll() throws Exception {
+    public long countAll() throws Exception {
 
         Connection connection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        int count = 0;
+        long count = 0;
 
         try {
         	
@@ -755,7 +755,7 @@ public final class ComponentOrderDAOJDBC implements ComponentOrderDAO {
 
             if (resultSet.next()) {
             	
-                count = resultSet.getInt("VALUE");
+                count = resultSet.getLong("VALUE");
             }
         } 
         catch (SQLException e) {
