+++ SourceCode/Java/DAOAnatomyJavaLayer/src/daojdbc/DerivedPartOfFKDAOJDBC.java
@@ -410,7 +410,7 @@ public final class DerivedPartOfFKDAOJDBC implements DerivedPartOfFKDAO{
     /*
      * Returns total amount of rows in table.
      */
-    public int count(String searchTerm) throws Exception {
+    public long count(String searchTerm) throws Exception {
 
         //String searchWithWildCards = "%" + searchTerm + "%";
         String searchWithWildCards = "%" + searchTerm;
@@ -422,7 +422,7 @@ public final class DerivedPartOfFKDAOJDBC implements DerivedPartOfFKDAO{
         Connection connection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
-        int count = 0;
+        long count = 0;
 
         try {
         	
@@ -441,7 +441,7 @@ public final class DerivedPartOfFKDAOJDBC implements DerivedPartOfFKDAO{
 
             if (resultSet.next()) {
             	
-                count = resultSet.getInt("VALUE");
+                count = resultSet.getLong("VALUE");
             }
         } 
         catch (SQLException e) {
